Fix bug with Touch_Controller::scale()
Touch_Controller::scale() was using the engine_interface object,
and is called from Touch_Controller::Touch_Controller().

A Touch_Controller was also a member of Engine_Interface, which
meant that when engine_interface was created, it went to create
its Touch_Controller member, but then THAT wanted data from the
still-constructing engine_interface.